Created a proper Auto Potion StateMachine.
Changed item cooldown tracking to be generic. Subscribing to client item use packets so that we can automatically put them in a queue which is used when handling an item use response from the server. Added functions in Self to check if we're in a state to use items in general and also for specific items.Created a proper Auto Potion StateMachine.