Improving botting/training logic
Improved logic which decides what to do when Botting starts (townloop or training).
Handling when a buff is added to us. This is useful because there is a gap between when we successfully cast a skill and when the buff becomes active.
Added an event for when an item in our inventory is updated. We use this for when an item's durability is updated.
Filtering out item/skill descriptions (_DESC) and skill "study"s (_STUDY) from text data.
Updated Geometry interface to allow for checking if a point is inside it.
Changed skill set from that for a euro char to that for a spear nuker.
Added consistent blocks to event handling switch statement.