Send path to UI. improve skill use and movement.
Sending character's walking path to the UI.
Updated setChildStateMachine to no longer take a reference to the Bot, since the base class already holds that reference.
Moved training area a little.
Added debug mechanism for when CastSkill seems to get stuck.
Slight change in skill handling; we no longer consider it an error if a non-root skill has a cooldown. This is probably just weird inconsistency by joymax.
Fixed CastSkillOnEntity to no longer do any distance checks on the given destination.
Added attacks and buffs to list.
Changed random walk to be a Walking StateMachine (that can be interrupted).