Make pvp preparation more thorough.
Change all LOG statements which are for caught exceptions to be logged as ERROR level.
Change pvp preparation to ensure hp/mp are full and we have no statuses.
Change pvp preparation to wait for item and skill cooldowns to end.
Introduce an ItemUseSuccess event.
Update UseItem to utilize the ItemUseSuccess to understand that an item has been successfully used.